Added support for role to user mapping
Routes from roles to users are no longer hardcoded. They are instead
specified when instanciating a form.

There are two kinds of mapping:
- single users roles correspond to a one-to-one mapping (one user for
  the role)
- group roles correspond to a one-to-many mapping (several users can
  have the role)

This has required some refactoring...

TOFIX: for a group role, when a user submit a result, all users
participating in the group role will recieve the same succes/failure
notification. Others users should require another kind of notification
(and only when one of the users succeded).