Finer task validation messages
In the context of group roles, the user which completed the task and the
others will receive different messages. On the other hand a rejected task
submission is only notified to the use which submitted it.
However several interfaces with the same user will always receive the
same message.
Also if something goes wrong with the completion of a task, the user
receive a notification about it.