This is a new proposal for the bridges. This branch depends on the acceptance of the branch "another-proposal-about-controllers"
Thus, like that other branch, it has not been discussed
yet with Jean. Also it could be appropriate to move the
XWIKI dependent implementation of the API (eu.learnpad.core.impl)
in another project.

Currently there was added a dependency with
eu.learnpad.rest.utils.RestResource which is still in
the old project implementing an XWIKI solution for
the controllers.