[sim] Overhaul of the build system
The previous system to generate uber jar had several limitations, which
threatened to cause problems (and actually started to do so in the REST
branch).

The initial plan was to replace assembly with the shade plugin, but we
reached a limit in the number of files which can be contained into a jar
file(!).

The current solution is described here:
http://stackoverflow.com/questions/13021423/invalid-or-corrupt-jar-file-built-by-maven-shade-plugin
and here:
https://ath3nd.wordpress.com/2013/12/25/packaging-a-multimodule-maven-spring-app-in-a-standalone-jar/