Simulator: BPMN explorer improvements
The BPMN explorer now exposes more functionalities. It now tracks data
objects (DOs) used in I/O of usertasks, as well as "standalone" DOs
(that is DOs which are not in output of any task).

In order to do this, the initialization of the explorer has be modified
to be cleaner.

Also, all collections exposed by the explorer are now immutables.