Simulator: WebUI handler now sends score to users
The WebUI handler now sends users their current session score with each
new task.
To avoid additional msg exchange, the session score is send with the new
tasks description.
To accommodate this, a slight change in the order of the messages has to
be done (but the current client side code base was already able to
handle it with no change): the task descr is send *after* the user sent
a subscribe message, and not at the opening of the connection (as
before). Actually, this seems more sensible in a way...
Note that the user session score are not yet displayed on the client
side.