Simulator: correct bug in proc. manager/dispatch
There was possible bug where a process dispatcher starts to send tasks
which are responded to *before* the manager had time to finish
registering it properly. This has been fixed, but sadly the dispatchers
now need to be started manually (which is not so bad as we already had
similar problems internally with abstract/implementation
initialization).
Tests have been updated accordingly.