Simulator: replace processDefId with processDefKey
Previous version of API used processDefId to refer to a process
definition. This was not good as it exposed outside to
activiti-specifics.
Instead, process definitions are now refered by the processDefKey, which
correspond to the Id defined in the BPMN. This make interfacing with the
simulator easier.
This required some new additional methods to translate process
definition keys to process definition IDs.