[mon] correct build script install deps. error
The build script contained a mistake when trying to install local jar
dependencies. It tried to get the base name of a jar using "tr -d .jar",
but since each character in ".jar" is handled separately it had the
infortunate effect of removing all the '.',  'j', 'a' and 'r' from the
file name.
Replace this part of the script with the `basename` command.