[sim] Add `integrated` flag to webui
Add a dedicated flag to indicate whether the web ui should work in
"standalone" or "integrated" mode.
In integrated mode many elements of the interface are disabled or work
differently than in the standalone mode.
To decide whether to work in standalone or integrated mode, the UI
checks if it is integrated in an iFrame (integrated) or not
(standalone).