[sim] Change Activiti event handling for tasks
Massive changes in preparation of Activiti message workaround.
Sadly, it is quite difficult to cut this in smaller commit, so it will
have to do.

This includes a rework on the way Activiti events are handled, that
avoid a *lot* of bugs/weirds behaviors:
The ActivitiProcessManager is now the sole listener for activiti
events. It handle them using a single-thread executor, where new events
are added to a job queue. This ensures that an event is completely
handled before trying to handle new events it may have triggered
(something that Activiti does not like *at all*).

Consequently the Activiti ProcessDispatcher no longer handle these
events in a synchrone fashion, something that simplify quite a lot of
things.