[sim] Change task validation/completion sequence
This commit introduces a major change in the way the tasks are
validated. Previous design was not thread-safe and could cause locking
under stress.

With the new design, the validation and completion of a task have been
separated, and thread-safety improved. This required a few internal API
changes.

Dispatcher Unit Testing was quite a huge mess. It has been
heavily refactored. Also a new test has been added to check for correct
task validation calls by a dispatcher.