Correct Sim user infos API path annotations
The path annotations on the IUserInfosAPI interface were incorrect. A
base path was set at the interface level, and relative path at the
methods level. This does not works as the interface is implemented by a
class with its own base path.

The annotations are now corrected to be at the methods level only.